requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-filter
markdown==3.6
django-filter==24.2
django-model-utils==4.5.1
django-environ~=0.8.1
psycopg2==2.9.9